Refactor de ubicación de la interface de observer y observado
En este ejemplo tenemos una app que tiene un unico MVC, pero en una
app grande, podemos tener muchos MVC, incluso un C que tiene a su
cargo muchos M y V diferentes.

Imaginemos que todos o alguno de ellos implementan el patron
Observer, y deberiamos ir a buscarlo al directorio del primer MVC
que construimos.

Mejor tenerlo en un dir especifico de cosas comunes.
